+++ src/static_dependency_injector/containers/__init__.py
@@ -1,3 +1,6 @@
 from static_dependency_injector.containers._static_declarative_container import StaticDeclarativeContainer
+from static_dependency_injector.containers._static_declarative_container_meta import (
+    UnannotatedProviderWarning,
+)
 
-__all__ = ("StaticDeclarativeContainer",)
+__all__ = ("StaticDeclarativeContainer", "UnannotatedProviderWarning")